# Bambi.Services/Dtos/Users/UpdateProfileDto.cs
namespace Bambi.Services.Dtos.Users;

public class UpdateProfileDto
{
    public string? Description { get; set; }
    public string? City { get; set; }
    public string? PhoneNumber { get; set; }
    public string? Email { get; set; }
}