# bambi-api-openapi.yaml
openapi: 3.0.0
info:
  title: Bambi - Second-Hand Clothing Marketplace API
  description: |
    Complete REST API for Bambi second-hand clothing marketplace platform.
    Includes user authentication, listings management, categories, purchases, and admin features.

    ### Enum serialization
    All enums (PurchaseStatus, UserRole, ItemGender) are serialised as **strings by name** in responses,
    and accept **both string names and integer values** in requests. e.g. either `"status": "Confirmed"`
    or `"status": 1` is valid.

    ### Validation
    Input is validated by **FluentValidation** before reaching the controller. Validation failures return
    `400 Bad Request` with an RFC 7807 ProblemDetails body, including per-field error messages under
    the `errors` extension.
  version: 1.1.0
  contact:
    name: Bambi Team
    url: https://bambi.local
  license:
    name: MIT
servers:
  - url: http://localhost:5138
    description: Development Server
  - url: https://api.bambi.local
    description: Production Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for authenticated endpoints
  schemas:
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          description: User unique identifier
          example: 1
        username:
          type: string
          maxLength: 50
          description: User's username
          example: "john_doe"
        email:
          type: string
          format: email
          maxLength: 150
          description: User's email address
          example: "john@example.com"
        phoneNumber:
          type: string
          maxLength: 20
          nullable: true
          description: User's phone number
          example: "+359888123456"
        city:
          type: string
          maxLength: 100
          nullable: true
          description: User's city
          example: "Sofia"
        description:
          type: string
          maxLength: 500
          nullable: true
          description: User-supplied bio / about-me text
          example: "Vintage hunter from Sofia. Ships within 48h."
        profilePicUrl:
          type: string
          format: uri
          nullable: true
          maxLength: 500
          description: URL to profile picture (Cloudinary)
          example: "https://res.cloudinary.com/bambi/image/upload/v1234567890/profile_pic.jpg"
        role:
          type: string
          enum: [user, admin]
          description: User's role in the system (lowercase)
          example: "user"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2026-05-24T10:00:00Z"
    UpdateRoleDto:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: [User, Admin, user, admin]
          description: |
            New role to assign to user. Accepts either casing — backend
            parses case-insensitively and stores as the enum value.
          example: "admin"
    UpdateProfileDto:
      type: object
      description: Partial-update DTO for the current user's own profile. All fields optional.
      properties:
        description:
          type: string
          maxLength: 500
          nullable: true
          description: New bio
          example: "Updated bio"
        city:
          type: string
          maxLength: 100
          nullable: true
          description: New city
          example: "Plovdiv"
        phoneNumber:
          type: string
          maxLength: 20
          nullable: true
          description: New phone number
          example: "+359888000111"
        email:
          type: string
          format: email
          maxLength: 150
          nullable: true
          description: New email address
          example: "new@example.com"

    # ==================== Category Schemas ====================
...
    PagedResponse:
      type: object
      properties:
        items:
          type: array
          description: List of items
          items: {}
        totalCount:
          type: integer
          description: Total number of items
          example: 150
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Items per page
          example: 12
        totalPages:
          type: integer
          description: Total number of pages
          example: 13

    # ==================== Error Schemas ====================
    # All error responses follow RFC 7807 ProblemDetails (application/problem+json),
    # produced by the global exception middleware. Validation failures raised by
    # FluentValidation include per-field messages under `errors`.
    ErrorResponse:
      type: object
      description: RFC 7807 ProblemDetails payload returned for all non-2xx responses.
      properties:
        type:
          type: string
          format: uri
          description: Reference URI for the error type
          example: "https://httpstatuses.com/400"
        title:
          type: string
          description: Short human-readable summary
          example: "Validation Failed"
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Detailed error message (for AppException-derived errors)
          example: "One or more validation errors occurred."
        instance:
          type: string
          description: Request path
          example: "/api/auth/register"
        errors:
          type: object
          nullable: true
          additionalProperties:
            type: array
            items:
              type: string
          description: |
            Field-level validation errors keyed by property name. Present only on
            FluentValidation failures (HTTP 400).
          example:
            Username: ["Username is required."]
            Email: ["Email is not a valid address."]
security:
  - BearerAuth: []
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: List all users
      description: |
        Retrieves paginated list of users (admin only).
        Used for user management in admin panel.
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Items per page
        - name: search
          in: query
          schema:
            type: string
          description: Search by username or email
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/UserDto'
        '401':
          description: Unauthorized - authentication required
        '403':
          description: Forbidden - admin role required
        '500':
          description: Internal server error
  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: |
        Retrieves user profile information.
        Returns public info for other users, full info for self (admin/user).
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: User not found
        '500':
          description: Internal server error

    put:
      tags:
        - Users
      summary: Update user profile
      description: |
        Updates a user's profile (bio, city, phone, email).
        Callable by the user themself **or** an admin. All fields are optional —
        only the keys present in the payload get applied.
      operationId: updateUserProfile
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Unauthorized — authentication required
        '403':
          description: Forbidden — only the user themself or an admin may update
        '404':
          description: User not found

    delete:
      tags:
        - Users
      summary: Delete user account
      description: |
        Deletes a user account and all associated data (admin only).
        Also deletes user's listings and purchase history.
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized - authentication required
        '403':
          description: Forbidden - admin role required
        '404':
          description: User not found
        '500':
          description: Internal server error
  /api/users/{id}/role:
    put:
      tags:
        - Users
      summary: Update user role
      description: |
        Changes user's role (admin only).
        Can promote/demote between user, seller, and admin roles.
      operationId: updateUserRole
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleDto'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Validation failed
        '401':
          description: Unauthorized - authentication required
        '403':
          description: Forbidden - admin role required
        '404':
          description: User not found
        '500':
          description: Internal server error
  /api/users/me/profile-pic:
    post:
      tags:
        - Users
      summary: Upload / replace the caller's profile picture
      description: |
        Uploads an image to Cloudinary and stores its URL on the current user.
        If the user already had a profile picture, the old one is deleted from
        Cloudinary before the new one is saved.
      operationId: uploadProfileImage
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (jpg, png, webp). Max 10 MB.
      responses:
        '200':
          description: Profile picture replaced. Returns the updated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '400':
          description: Missing or invalid file
        '401':
          description: Unauthorized — authentication required
  /api/users/me:
    delete:
      tags:
        - Users
      summary: Delete own account
      description: |
        Deletes the current user's account and all associated data.
        
        Validations:
        - User cannot have active purchases as buyer (pending, confirmed, or shipped)
        - User cannot have active sales (pending, confirmed, or shipped purchases)
        - Deletes all user's listings and images
        - Deletes all user's purchase records
        
        Side Effects:
        - Permanently removes account
        - All listings become unavailable
        - All images deleted from storage
        - Cannot be undone
      operationId: deleteMyAccount
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Account deleted successfully
        '401':
          description: Unauthorized - authentication required
        '409':
          description: Conflict - cannot delete with active purchases or sales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error

  # ==================== PURCHASE ENDPOINTS ====================
...